04_KDP_Automation/requirements.txt
anthropic
ebooklib
Pillow
google-generativeai
